Refactored the move nodes

+++ doubly_linked_list/doubly_linked_list.py
@@ -110,7 +110,7 @@ class DoublyLinkedList:
     def remove_from_tail(self):
         if not self.head and not self.tail:
             return
-        
+
         removed_node_value = self.tail.value
 
         if self.head == self.tail:
@@ -123,20 +123,39 @@ class DoublyLinkedList:
         self.length -= 1
         return removed_node_value
 
-
     """Removes the input node from its current spot in the 
     List and inserts it as the new head node of the List."""
 
     def move_to_front(self, node):
-        self.add_to_head(node.value)
-        node.delete()
-        self.length -= 1
+        if node == self.head:
+            return
+
+        node_value = node.value
+
+        if node == self.tail:
+            self.remove_from_tail()
+        else:
+            node.delete()
+            self.length -= 1
+
+        self.add_to_head(node_value)
 
     """Removes the input node from its current spot in the 
     List and inserts it as the new tail node of the List."""
 
     def move_to_end(self, node):
-        pass
+        if node == self.tail:
+            return
+
+        node_value = node.value
+
+        if node == self.head:
+            self.remove_from_head()
+        else:
+            node.delete()
+            self.length -= 1
+
+        self.add_to_tail(node_value)
 
     """Removes a node from the list and handles cases where
     the node was the head or the tail"""
